Allow a user to be an admin of more than one trac instance
This adds some hacky code that gets around puppet's current lack
of iteration support.  It is especially trickey since the iteration
is being done insiade a defined type, which means the same iteration
loop could be required multiple times.  So if it looks strange, even
to those familiar with the define loop pattern, that's why.  I tried
to explain it in the code comments.